Moved Python bindings of interfaces into separate files; created separate file for Python utilities
