+++ reach/src/python/python_bindings.cpp
@@ -13,7 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <reach_core/reach_study.h>
+#include <reach/reach_study.h>
+#include "display.hpp"
+#include "evaluator.hpp"
+#include "ik_solver.hpp"
+#include "logger.hpp"
+#include "target_pose_generator.hpp"
+#include "utils.h"
 
 #include <boost_plugin_loader/plugin_loader.hpp>
 #include <boost/python.hpp>
@@ -26,304 +32,6 @@ namespace bp = boost::python;
 
 namespace reach
 {
-void print_py_error(std::stringstream& ss)
-{
-  try
-  {
-    PyErr_Print();
-    bp::object sys(bp::handle<>(PyImport_ImportModule("sys")));
-    bp::object err = sys.attr("stderr");
-    std::string err_text = bp::extract<std::string>(err.attr("getvalue")());
-    ss << err_text << std::endl;
-  }
-  catch (...)
-  {
-    ss << "Failed to parse python error" << std::endl;
-  }
-  PyErr_Clear();
-}
-
-template <typename Function, typename ClassT, typename... Args>
-auto call_and_handle(Function func, ClassT* obj, const char* function_name, Args&&... args)
-{
-  try
-  {
-    return (obj->*func)(std::forward<Args>(args)...);
-  }
-  catch (const bp::error_already_set& e)
-  {
-    std::stringstream ss;
-    ss << "Python exception in " << function_name << ": ";
-    print_py_error(ss);
-
-    std::runtime_error exception(ss.str());
-    throw exception;
-  }
-}
-
-struct IKSolverPython : IKSolver, bp::wrapper<IKSolver>
-{
-  std::vector<std::string> getJointNamesFunc() const
-  {
-    std::vector<std::string> names;
-
-    bp::list name_list = this->get_override("getJointNames")();
-
-    for (int i = 0; i < bp::len(name_list); ++i)
-    {
-      std::string name = bp::extract<std::string>(name_list[i]);
-      names.push_back(name);
-    }
-
-    return names;
-  }
-  std::vector<std::string> getJointNames() const
-  {
-    return call_and_handle(&IKSolverPython::getJointNamesFunc, this, "getJointNames()");
-  }
-
-  std::vector<std::vector<double>> solveIKFunc(const Eigen::Isometry3d& target,
-                                               const std::map<std::string, double>& seed) const
-  {
-    std::vector<std::vector<double>> output;
-
-    bp::tuple shape = bp::make_tuple(4, 4);
-    bp::numpy::dtype dtype = bp::numpy::dtype::get_builtin<double>();
-    bp::numpy::ndarray array = bp::numpy::zeros(shape, dtype);
-
-    for (int i = 0; i < 4; ++i)
-    {
-      for (int j = 0; j < 4; ++j)
-      {
-        array[i, j] = target.matrix()(i, j);
-      }
-    }
-
-    bp::dict dictionary;
-    for (std::pair<std::string, double> pair : seed)
-    {
-      dictionary[pair.first] = pair.second;
-    }
-
-    bp::list list = this->get_override("solveIK")(array, dictionary);
-
-    for (int i = 0; i < bp::len(list); ++i)
-    {
-      std::vector<double> sub_vec;
-      bp::list sub_list = bp::extract<bp::list>(list[i]);
-      for (int j = 0; j < bp::len(sub_list); ++j)
-      {
-        sub_vec.push_back(bp::extract<double>(sub_list[j]));
-      }
-      output.push_back(sub_vec);
-    }
-
-    return output;
-  }
-  std::vector<std::vector<double>> solveIK(const Eigen::Isometry3d& target,
-                                           const std::map<std::string, double>& seed) const
-  {
-    return call_and_handle(&IKSolverPython::solveIKFunc, this, "solveIK()", target, seed);
-  }
-};
-
-struct IKSolverFactoryPython : IKSolverFactory, bp::wrapper<IKSolverFactory>
-{
-  IKSolver::ConstPtr createFunc(const YAML::Node& config) const
-  {
-    return this->get_override("create")(config);
-  }
-  IKSolver::ConstPtr create(const YAML::Node& config) const
-  {
-    return call_and_handle(&IKSolverFactoryPython::createFunc, this, "IKSolverFactory::create", config);
-  }
-};
-
-struct EvaluatorPython : Evaluator, bp::wrapper<Evaluator>
-{
-  double calculateScoreFunc(const std::map<std::string, double>& map) const
-  {
-    bp::dict dictionary;
-    for (auto pair : map)
-    {
-      dictionary[pair.first] = pair.second;
-    }
-
-    double score = this->get_override("calculateScore")(dictionary);
-
-    return score;
-  }
-  double calculateScore(const std::map<std::string, double>& map) const
-  {
-    return call_and_handle(&EvaluatorPython::calculateScoreFunc, this, "calculateScore()", map);
-  }
-};
-
-struct EvaluatorFactoryPython : EvaluatorFactory, bp::wrapper<EvaluatorFactory>
-{
-  Evaluator::ConstPtr createFunc(const YAML::Node& config) const
-  {
-    return this->get_override("create")(config);
-  }
-  Evaluator::ConstPtr create(const YAML::Node& config) const
-  {
-    return call_and_handle(&EvaluatorFactoryPython::createFunc, this, "EvaluatorFactory::create", config);
-  }
-};
-
-struct TargetPoseGeneratorPython : TargetPoseGenerator, bp::wrapper<TargetPoseGenerator>
-{
-  VectorIsometry3d generateFunc() const
-  {
-    VectorIsometry3d eigen_list;
-
-    bp::list np_list = this->get_override("generate")();
-
-    // Convert the list of 4x4 numpy arrays to VectorIsometry3d
-    for (int i = 0; i < bp::len(np_list); ++i)
-    {
-      Eigen::Isometry3d eigen_mat;
-      bp::numpy::ndarray np_array = bp::numpy::from_object(np_list[i]);
-
-      for (int j = 0; j < 4; ++j)
-      {
-        for (int k = 0; k < 4; ++k)
-        {
-          eigen_mat.matrix()(j, k) = bp::extract<double>(np_array[j][k]);
-        }
-      }
-      eigen_list.push_back(eigen_mat);
-    }
-
-    return eigen_list;
-  }
-  VectorIsometry3d generate() const
-  {
-    return call_and_handle(&TargetPoseGeneratorPython::generateFunc, this, "generate()");
-  }
-};
-
-struct TargetPoseGeneratorFactoryPython : TargetPoseGeneratorFactory, bp::wrapper<TargetPoseGeneratorFactory>
-{
-  TargetPoseGenerator::ConstPtr createFunc(const YAML::Node& config) const
-  {
-    return this->get_override("create")(config);
-  }
-  TargetPoseGenerator::ConstPtr create(const YAML::Node& config) const
-  {
-    return call_and_handle(&TargetPoseGeneratorFactoryPython::createFunc, this, "TargetPoseGeneratorFactory::create",
-                           config);
-  }
-};
-
-struct DisplayPython : Display, bp::wrapper<Display>
-{
-  void showEnvironmentFunc() const
-  {
-    this->get_override("showEnvironment")();
-  }
-  void showEnvironment() const
-  {
-    return call_and_handle(&DisplayPython::showEnvironmentFunc, this, "showEnvironment()");
-  }
-
-  void updateRobotPoseFunc(const std::map<std::string, double>& map) const
-  {
-    bp::dict dictionary;
-    for (auto pair : map)
-    {
-      dictionary[pair.first] = pair.second;
-    }
-
-    this->get_override("updateRobotPose")(dictionary);
-  }
-  void updateRobotPose(const std::map<std::string, double>& map) const
-  {
-    return call_and_handle(&DisplayPython::updateRobotPoseFunc, this, "updateRobotPose()", map);
-  }
-
-  void showReachNeighborhoodFunc(const std::vector<ReachRecord>& neighborhood) const
-  {
-    this->get_override("showReachNeighborhood")(neighborhood);
-  }
-  void showReachNeighborhood(const std::vector<ReachRecord>& neighborhood) const
-  {
-    return call_and_handle(&DisplayPython::showReachNeighborhoodFunc, this, "showReachNeighborhood()", neighborhood);
-  }
-
-  void showResultsFunc(const ReachDatabase& results) const
-  {
-    this->get_override("showResults")(results);
-  }
-  void showResults(const ReachDatabase& results) const
-  {
-    return call_and_handle(&DisplayPython::showResultsFunc, this, "showResults()", results);
-  }
-};
-
-struct DisplayFactoryPython : DisplayFactory, bp::wrapper<DisplayFactory>
-{
-  Display::ConstPtr createFunc(const YAML::Node& config) const
-  {
-    return this->get_override("create")(config);
-  }
-  Display::ConstPtr create(const YAML::Node& config) const
-  {
-    return call_and_handle(&DisplayFactoryPython::createFunc, this, "DisplayFactory::create", config);
-  }
-};
-
-struct LoggerPython : Logger, bp::wrapper<Logger>
-{
-  void setMaxProgressFunc(unsigned long max_progress)
-  {
-    this->get_override("setMaxProgress")(max_progress);
-  }
-  void setMaxProgress(unsigned long max_progress)
-  {
-    return call_and_handle(&LoggerPython::setMaxProgressFunc, this, "setMaxProgress()", max_progress);
-  }
-
-  void printProgressFunc(unsigned long progress) const
-  {
-    this->get_override("printProgress")(progress);
-  }
-  void printProgress(unsigned long progress) const
-  {
-    return call_and_handle(&LoggerPython::printProgressFunc, this, "printProgress()", progress);
-  }
-
-  void printResultsFunc(const StudyResults& results) const
-  {
-    this->get_override("printResults")(results);
-  }
-  void printResults(const StudyResults& results) const
-  {
-    return call_and_handle(&LoggerPython::printResultsFunc, this, "printResults()", results);
-  }
-
-  void printFunc(const std::string& msg) const
-  {
-    this->get_override("print")(msg);
-  }
-  void print(const std::string& msg) const
-  {
-    return call_and_handle(&LoggerPython::printFunc, this, "print()", msg);
-  }
-};
-
-struct LoggerFactoryPython : LoggerFactory, bp::wrapper<LoggerFactory>
-{
-  Logger::Ptr createFunc(const YAML::Node& config) const
-  {
-    return this->get_override("create")(config);
-  }
-  Logger::Ptr create(const YAML::Node& config) const
-  {
-    return call_and_handle(&LoggerFactoryPython::createFunc, this, "LoggerFactory::create", config);
-  }
-};
-
 class ReachStudyPython : public ReachStudy
 {
 public:
@@ -530,4 +238,5 @@ BOOST_PYTHON_MODULE(reach_core_python)
     bp::register_ptr_to_python<TargetPoseGeneratorFactory::Ptr>();
   }
 }
+
 }  // namespace reach
